refactor(cli): SMI-5127 — withTelemetry sibling-split pilot (sync, search)
Establishes the <command>.action.ts sibling-split convention for the SMI-5083
telemetry long-tail: sync.ts and search.ts exceed the 500-line gate once their
action impls are wrapped, so the impls move to sync.action.ts / search.action.ts
and the command factories import the wrapped actions.

Wraps all 5 handlers with withTelemetry (source: 'cli'): syncAction +
sync{Status,History,Config}Action (namespaced skillIds 'sync', 'sync status',
'sync history', 'sync config') and searchAction. Extends CLI_DISPATCHER_MAP with
the sibling-base keys 'sync.action' and 'search.action'.

Post-split lengths: sync.ts 108, sync.action.ts 439, search.ts 72,
search.action.ts 461 (all <500). Coverage test green (9 dispatchers); tsc/eslint/
prettier clean; no consumer imports broken (search.ts re-exports preserved).
